Add files via upload
Code and tools to build an Arduino Lipo Battery monitor. VoltageDivider is a spreadsheet that shows all fractions given a set of Resistors. The Schematic shows how to wire. the sketch_LipoBattery.ino contains code to download to the Arduino and library/RunningStat allows on-the-fly statistics calculation for mean and standardDeviation.  Place the RunningStat files in a directory RunningStat under the Arduino/libraries.